Refactor: 상태 확인 컴포넌트 분리 #71
