+++ src/pages/mypage/reservations/[id]/index.tsx
@@ -1,18 +1,18 @@
 import Image from 'next/image';
-import { ReactElement, useState } from 'react';
 import Link from 'next/link';
+import { ReactElement, useState } from 'react';
 
 import ArrowIcon from '#/icons/icon-arrowBack.svg';
-import ArrowRightIcon from '#/icons/icon-arrowRight.svg';
 import AlertModal from '@/components/Modal/AlertModal/AlertModal';
 import ReviewModal from '@/components/Modal/ReviewModal/ReviewModal';
 import MyPageLayout from '@/components/MyPage/MyPageLayout';
+import CheckStatus from '@/components/Reservations/Id/CheckStatus';
 import Button from '@/components/common/Button/Button';
-import { CANCELED, COMPLETED, CONFIRMED, DECLINED, PENDING, RESERVATION_STATUS } from '@/constants/reservation';
-import { priceFormat } from '@/utils/priceFormat';
-import styles from './ReservationId.module.css';
-import dayjs from 'dayjs';
+import { COMPLETED, PENDING } from '@/constants/reservation';
 import { Reservation } from '@/types/common/api';
+import { priceFormat } from '@/utils/priceFormat';
+import dayjs from 'dayjs';
+import styles from './ReservationId.module.css';
 
 // @todo id값을 사용해서 해당 카드 데이터 불러오기
 const item: Reservation = {
@@ -38,32 +38,6 @@ const item: Reservation = {
 
 // @todo 주소 데이터 불러오기
 const address = '서울특별시 강남구 테헤란로 427';
-
-interface CheckStatusProps {
-  status: string;
-}
-
-const CheckStatus = ({ status }: CheckStatusProps) => {
-  return (
-    <div className={styles.status}>
-      {status === CANCELED || status === DECLINED ? (
-        <div className={styles[status]}>{RESERVATION_STATUS[status]}</div>
-      ) : (
-        <>
-          <div className={status === PENDING ? styles.active : styles.inactive}>{RESERVATION_STATUS['pending']}</div>
-          <ArrowRightIcon alt="예약 신청에서 예약 승인으로 가는 화살표" />
-          <div className={status === CONFIRMED ? styles.active : styles.inactive}>
-            {RESERVATION_STATUS['confirmed']}
-          </div>
-          <ArrowRightIcon alt="예약 승인에서 체험 완료로 가는 화살표" />
-          <div className={status === COMPLETED ? styles.active : styles.inactive}>
-            {RESERVATION_STATUS['completed']}
-          </div>
-        </>
-      )}
-    </div>
-  );
-};
 
 function ReservationID() {
   const [isAlertModalOpen, setIsAlertModalOpen] = useState(false);
